requirements.txt
backcall==0.1.0
bleach==3.0.2
blessings==1.7
bokeh==0.12.16
boto3==1.10.39
botocore==1.13.39
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
cycler==0.10.0
Cython==0.28.3
cytoolz==0.9.0.1
dask==0.18.0
decorator==4.3.0
defusedxml==0.5.0
distributed==1.22.0
docutils==0.15.2
dominate==2.3.5
entrypoints==0.2.3
future==0.17.1
gpustat==0.5.0
heapdict==1.0.0
idna==2.8
imageio==2.3.0
imgaug==0.2.6
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.4
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
locket==0.2.0
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.4
mkl-fft==1.0.0
mkl-random==1.0.1
msgpack-python==0.5.6
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
notebook==5.7.2
numpy==1.14.5
nvidia-ml-py3==7.352.0
olefile==0.45.1
opencv-python==3.4.2.17
openslide-python==1.1.1
packaging==17.1
pandas==0.23.1
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.0
psutil==5.4.6
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.3.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytorch-pretrained-biggan==0.1.1
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
s3transfer==0.2.1
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.11.0
sortedcontainers==2.0.4
spams==2.6.1
staintools==2.0.1
tblib==1.3.2
tensorboardX==1.2
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==0.4.1.post2
torchfile==0.1.0
torchvision==0.2.1
tornado==5.0.2
tqdm==4.24.0
traitlets==4.3.2
urllib3==1.24.1
visdom==0.1.8.8
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
widgetsnbextension==3.4.2
zict==0.1.3
